Got rid of the ugly offset on the right of the MultiChannelLoudnessBar.

+++ projects/LUFSMeter/Source/gui/MultiChannelLoudnessBar.h
@@ -78,7 +78,7 @@ private:
     
     float stretch;
     float offset;
-    float widthOfIndividualChannel;
+    int widthOfIndividualChannel;
     
     Colour colour;
